Fix number of Codacy & LGTM warnings, update Pillow version (security alert).

+++ pywinauto/linux/application.py
@@ -120,7 +120,7 @@ class Application(BaseApplication):
         except Exception:
             raise ProcessNotFoundError()
 
-    def kill(self):
+    def kill(self, soft=False):
         """
         Try to close and kill the application
 
@@ -129,7 +129,7 @@ class Application(BaseApplication):
         This should only be used when it is OK to kill the process like you
         would do in task manager.
         """
-        if self._proc_descriptor is not None:
+        if soft and self._proc_descriptor is not None:
             # Kill process created via Application with subprocess kill
             self._proc_descriptor.kill()
             # wait for child process to terminate
@@ -137,11 +137,13 @@ class Application(BaseApplication):
             self._proc_descriptor = None
 
         if not self.is_process_running():
+            self._proc_descriptor = None
             return True # already closed
         status = subprocess.check_output(["kill", "-9", str(self.process)], universal_newlines=True)
         if "Operation not permitted" in status:
             raise Exception("Cannot kill process: {}".format(status))
         else:
+            self._proc_descriptor = None
             return True
 
     def is_process_running(self):
